+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * 【db     】     【不向下兼容】增加MongoDB4.x支持（pr#568@Gitee）
 * 【json   】     【可能兼容问题】修改JSONObject结构，继承自MapWrapper
 * 【core   】     【可能兼容问题】BeanCopier重构，新建XXXCopier，删除XXXValueProvider
+* 【core   】     【可能兼容问题】URLEncoder废弃，URLEncoderUtil使用RFC3986
 
 ### 🐣新特性
 * 【http   】     HttpRequest.form采用TableMap方式（issue#I4W427@Gitee）